+++ src/error_metrics/error_rate.cpp
@@ -55,7 +55,7 @@ namespace abo::error_metrics {
 
     double error_rate_add(const Cudd &mgr, const std::vector<BDD> &f, const std::vector<BDD> &f_hat) {
         ADD diff = abo::util::xor_difference_add(mgr, f, f_hat);
-        std::vector<std::pair<unsigned long, unsigned long>> terminal_values = abo::util::add_terminal_values(diff);
+        std::vector<std::pair<double, unsigned long>> terminal_values = abo::util::add_terminal_values(diff);
         unsigned long non_zero_path_count = 0;
         unsigned long total_path_count = 0;
         for (auto [value, path_count] : terminal_values) {
