+++ src/util/cudd_helpers.hpp
@@ -24,7 +24,7 @@ namespace abo::util {
     // return the value of the given add node, it must be a constant
     unsigned int const_add_value(const ADD &add);
     // returns a list of values (first) and how often they are reached (second)
-    std::vector<std::pair<unsigned long, unsigned long> > add_terminal_values(const ADD &add);
+    std::vector<std::pair<double, unsigned long> > add_terminal_values(const ADD &add);
 
     // assumes that the function represents an unsigned integer
     ADD bdd_forest_to_add(const Cudd &mgr, const std::vector<BDD> &bdds);
